refactor: change folder structure and paths
