+++ package.json
@@ -66,15 +66,19 @@
     "compile": "tsc -p ./",
     "watch": "tsc -watch -p ./",
     "pretest": "pnpm run compile && pnpm run lint",
-    "lint": "eslint src --ext ts",
+    "lint": "eslint \"./**/*.{ts,tsx}\" --fix",
+    "type-check": "tsc --noEmit",
+    "test": "echo 'no tests, yet... 😈",
     "format": "prettier --write \"**/*.{ts,tsx,md}\"",
     "prepare": "husky install",
+    "pre-commit": "pretty-quick --staged",
     "clear:node-modules": "npx rimraf ./**/node_modules",
     "clear:builds": "npx rimraf ./**/.next && npx rimraf ./**/.expo",
     "clear:project": "pnpm clear:node-modules && pnpm clear:builds",
     "clear:project-and-install": "pnpm clear:project && pnpm"
   },
   "devDependencies": {
+    "@changesets/cli": "^2.26.0",
     "@commitlint/cli": "^17.5.0",
     "@commitlint/config-conventional": "^17.4.4",
     "@types/glob": "^8.1.0",
@@ -84,13 +88,19 @@
     "@typescript-eslint/eslint-plugin": "^5.53.0",
     "@typescript-eslint/parser": "^5.53.0",
     "@vscode/test-electron": "^2.2.3",
+    "commitlint-plugin-workspace-scopes": "^1.1.0",
     "eslint": "^8.34.0",
-    "husky": "^8.0.0",
-    "lint-staged": "^13.0.1",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-config-turbo": "latest",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-unused-imports": "^2.0.0",
+    "husky": "^8.0.3",
+    "lint-staged": "^13.2.0",
     "mocha": "^10.2.0",
-    "prettier": "^2.8.3",
+    "prettier": "^2.8.4",
     "pretty-quick": "^3.1.3",
-    "rimraf": "^3.0.2",
+    "rimraf": "^4.1.2",
     "typescript": "^4.9.5"
   },
   "dependencies": {
