+++ src/commands/scan-file.command.ts
@@ -1,0 +1,20 @@
+import * as vscode from 'vscode';
+import { scanFiles } from '../utils/sdk';
+
+/**
+ * Scans the currently open file and highlights any errors found
+ */
+export const scanFileCommand = vscode.commands.registerCommand(
+  'extension.scanFileSdk',
+  () => {
+    const editor = vscode.window.activeTextEditor;
+
+    if (!editor) {
+      vscode.window.showErrorMessage('No active editor found.');
+      return;
+    }
+
+    const filePath = editor.document.uri.fsPath;
+    scanFiles([filePath], true);
+  }
+);
