+++ src/ui/buttons.status-bar.ts
